Revert "dataset modal routing and snapshot changelog"
