+++ app/src/sass/crn_app/_common.papaya.scss
@@ -156,8 +156,7 @@
     position:absolute;
     border:solid 2px darkgray;
     padding:4px;
-    margin:0;
-    left: 0px !important;
+    margin:0
 }
 .papaya-menu li {
     font-size:12px;
