+++ app/src/sass/crn_app/_dataset.dataset.scss
@@ -91,31 +91,5 @@
             }
         }
 
-        .dataset-form {
-            .dataset-form-header {
-                label {
-                    font-size: 18px;
-                }
-            }
-            .dataset-form-controls {
-                margin-top: 20px;
-            }
-
-            .changelog-length-warning, .btn-dataset-submit {
-                float: right;
-                margin-right: 20px;
-            }
-
-            .change {
-                margin-top: 10px;
-            }
-
-            .change-controls {
-                i {
-                    margin-right: 3px;
-                }
-            }
-        }
-
     }
 }
