+++ app/src/scripts/dataset/dataset.actions.js
@@ -7,7 +7,6 @@ var Actions = Reflux.createActions([
   'createSnapshot',
   'createComment',
   'cancelDirectoryUpload',
-  'constructDatasetUrl',
   'createSubscription',
   'deleteAttachment',
   'deleteComment',
@@ -20,7 +19,6 @@ var Actions = Reflux.createActions([
   'dismissError',
   'dismissJobsModal',
   'dismissMetadataIssue',
-  'dismissModals',
   'displayFile',
   'downloadLogs',
   'editFile',
@@ -44,7 +42,6 @@ var Actions = Reflux.createActions([
   'saveDescription',
   'selectJob',
   'setInitialState',
-  'showDatasetComponent',
   'sortComments',
   'startJob',
   'cancelJob',
@@ -56,7 +53,6 @@ var Actions = Reflux.createActions([
   'toggleSidebar',
   'uploadAttachment',
   'updateComment',
-  'updateCHANGES',
   'updateDescription',
   'updateName',
   'updateNote',
