+++ app/src/scripts/dataset/dataset.store.js
@@ -87,7 +87,6 @@ let datasetStore = Reflux.createStore({
       loadingApps: false,
       loadingJobs: false,
       loadingTree: false,
-      loadedUrl: null,
       jobs: [],
       metadataIssues: {},
       modals: {
@@ -98,7 +97,6 @@ let datasetStore = Reflux.createStore({
         share: false,
         update: false,
         subscribe: false,
-        snapshot: false,
       },
       redirectUrl: null,
       snapshot: false,
@@ -134,17 +132,6 @@ let datasetStore = Reflux.createStore({
     options = options ? options : {}
     options.isPublic = !userStore.data.token
 
-    // don't reload the current dataset
-    if (!forceReload && dataset && dataset._id === datasetId) {
-      this.update({ loading: false, loadingJobs: false })
-      return
-    }
-
-    // begin loading
-    // if (this.data.loading) {
-    //   return
-    // }
-
     // set active job if passed in query param
     if (options) {
       this.update({
@@ -156,15 +143,20 @@ let datasetStore = Reflux.createStore({
       })
     }
 
-    // update selection & current upload data, as well as loading state
+    // update selection & current upload data
     this.update({
       selectedSnapshot: datasetId,
       currentUploadId: uploadStore.data.projectId,
-      loading: true,
-      loadingJobs: true,
-      datasetTree: null,
-      loadedUrl: datasetId,
     })
+
+    // don't reload the current dataset
+    if (!forceReload && dataset && dataset._id === datasetId) {
+      this.update({ loading: false, loadingJobs: false })
+      return
+    }
+
+    // begin loading
+    this.update({ loading: true, loadingJobs: true, datasetTree: null })
     bids.getDataset(
       datasetId,
       res => {
@@ -182,33 +174,25 @@ let datasetStore = Reflux.createStore({
           let selectedSnapshot = this.data.selectedSnapshot
           if (!selectedSnapshot || selectedSnapshot === datasetId) {
             let dataset = res
+            this.update({
+              dataset,
+              datasetTree: [
+                {
+                  _id: datasetId,
+                  label: dataset.label,
+                  children: dataset.children,
+                },
+              ],
+            })
             let originalId = dataset.original ? dataset.original : datasetId
-            this.update(
-              {
-                dataset,
-                datasetTree: [
-                  {
-                    _id: datasetId,
-                    label: dataset.label,
-                    children: dataset.children,
-                  },
-                ],
-              },
-              this.loadJobs(datasetId, snapshot, originalId, options, jobs => {
-                this.loadSnapshots(dataset, jobs, () => {
-                  this.loadComments(originalId)
-                  this.checkUserSubscription(() => {
-                    let datasetUrl = this.constructDatasetUrl(dataset)
-                    this.update({
-                      loading: false,
-                      snapshot: snapshot,
-                      datasetUrl: datasetUrl,
-                    })
-                  })
+            this.loadJobs(datasetId, snapshot, originalId, options, jobs => {
+              this.loadSnapshots(dataset, jobs, () => {
+                this.loadComments(originalId)
+                this.checkUserSubscription(() => {
+                  this.update({ loading: false, snapshot: snapshot })
                 })
-              }),
-            )
-
+              })
+            })
             if (
               forceReload ||
               (!this.data.uploading && dataset.tags.includes('updating'))
@@ -529,6 +513,12 @@ let datasetStore = Reflux.createStore({
     modals[name] = !modals[name]
     update.modals = modals
 
+    // don't display the follow modal if the user is already following
+    if (name === 'subscribe' && this.data.dataset.subscribed) {
+      update.modals[name] = false
+    }
+    this.update(update)
+
     // callback
     if (callback && typeof callback === 'function') {
       callback()
@@ -554,20 +544,6 @@ let datasetStore = Reflux.createStore({
     this.update(update)
     if (callback && typeof callback === 'function') {
       callback()
-    }
-  },
-
-  showDatasetComponent(name, history, callback) {
-    let datasetUrl = this.data.datasetUrl
-    if (datasetUrl) {
-      let redirectUrl = datasetUrl
-      if (name !== '') {
-        redirectUrl = datasetUrl + '/' + name
-      }
-      history.push(redirectUrl)
-      if (callback) {
-        callback()
-      }
     }
   },
 
@@ -742,29 +718,6 @@ let datasetStore = Reflux.createStore({
       .then(res => {
         callback(null, res)
         dataset.README = value
-        this.update({ dataset })
-        this.updateModified()
-      })
-  },
-
-  /**
-   * Update CHANGES
-   */
-  updateCHANGES(value, callback) {
-    let dataset = this.data.dataset
-
-    scitran
-      .updateFileFromString(
-        'projects',
-        this.data.dataset._id,
-        'CHANGES',
-        value,
-        '',
-        [],
-      )
-      .then(res => {
-        callback(null, res)
-        dataset.CHANGES = value
         this.update({ dataset })
         this.updateModified()
       })
@@ -1183,14 +1136,14 @@ let datasetStore = Reflux.createStore({
           ) {
             this.updateDescriptionFile(file, this.data.dataset._id, () => {
               this.updateFileState(item, { loading: false })
-              this.showDatasetComponent('', item.history)
+              this.dismissModals()
             })
           } else {
             scitran
               .updateFile('projects', this.data.dataset._id, file)
               .then(() => {
                 this.updateFileState(item, { loading: false })
-                this.showDatasetComponent('', item.history)
+                this.dismissModals()
                 this.revalidate()
               })
           }
@@ -1675,8 +1628,10 @@ let datasetStore = Reflux.createStore({
   /**
    * DisplayFile
    */
-  displayFile(snapshotId, jobId, file, history, callback) {
+  displayFile(snapshotId, jobId, file, callback) {
     let requestAndDisplay = link => {
+      let modals = this.data.modals
+      modals.displayFile = true
       if (
         files.hasExtension(file.name, [
           '.pdf',
@@ -1690,20 +1645,19 @@ let datasetStore = Reflux.createStore({
         if (callback) {
           callback()
         }
-        this.showDatasetComponent('file-display', file.history)
         this.update({
           displayFile: {
             name: file.name,
             text: null,
             link: link,
           },
+          modals,
         })
       } else {
         request.get(link, {}).then(res => {
           if (callback) {
             callback()
           }
-          this.showDatasetComponent('file-display', file.history)
           this.update({
             displayFile: {
               name: file.name,
@@ -1711,6 +1665,7 @@ let datasetStore = Reflux.createStore({
               link: link,
               info: file,
             },
+            modals,
           })
         })
       }
@@ -1734,12 +1689,13 @@ let datasetStore = Reflux.createStore({
    */
   editFile(snapshotId, jobId, file, callback) {
     let requestAndDisplay = link => {
+      let modals = this.data.modals
+      modals.editFile = true
       if (files.hasExtension(file.name, ['.json', '.csv', '.tsv'])) {
         request.get(link, {}).then(res => {
           if (callback) {
             callback()
           }
-          this.showDatasetComponent('file-edit', file.history)
           this.update({
             editFile: {
               name: file.name,
@@ -1747,6 +1703,7 @@ let datasetStore = Reflux.createStore({
               link: link,
               info: file,
             },
+            modals,
           })
         })
       }
@@ -1874,19 +1831,6 @@ let datasetStore = Reflux.createStore({
         callback()
       }
     })
-  },
-
-  constructDatasetUrl(dataset) {
-    if (dataset) {
-      if (!dataset.original) {
-        let datasetId = dataset.linkID
-        return '/datasets/' + datasetId
-      } else {
-        let datasetId = dataset.linkOriginal
-        let version = dataset.linkID
-        return '/datasets/' + datasetId + '/versions/' + version
-      }
-    }
   },
 
   // usage analytics ---------------------------------------------------------------
