+++ app/src/scripts/dataset/tools/jobs/index.js
@@ -3,27 +3,22 @@
 // dependencies -------------------------------------------------------
 
 import React from 'react'
-import Reflux from 'reflux'
 import PropTypes from 'prop-types'
-import { Link, withRouter } from 'react-router-dom'
+import { withRouter } from 'react-router-dom'
 import actions from '../../dataset.actions.js'
-import datasetStore from '../../dataset.store.js'
 import Spinner from '../../../common/partials/spinner.jsx'
+import { Modal } from '../../../utils/modal.jsx'
 import moment from 'moment'
 import Results from '../../../upload/upload.validation-results.jsx'
 import Description from './description.jsx'
 import Parameters from './parameters.jsx'
-import Timeout from '../../../common/partials/timeout.jsx'
-import ErrorBoundary from '../../../errors/errorBoundary.jsx'
 import batch from '../../../utils/batch'
-import { refluxConnect } from '../../../utils/reflux'
 
-class JobMenu extends Reflux.Component {
+class JobMenu extends React.Component {
   // life cycle events --------------------------------------------------
 
   constructor() {
     super()
-    refluxConnect(this, datasetStore, 'datasets')
     this.state = {
       loading: false,
       parameters: {},
@@ -46,34 +41,27 @@ class JobMenu extends Reflux.Component {
   }
 
   componentWillReceiveProps() {
-    let datasets = this.state.datasets
-    let dataset = datasets ? datasets.dataset : null
-    let summary = dataset ? dataset.summary : null
-    let snapshots = datasets ? datasets.snapshots : null
-
-    if (dataset) {
-      // initialize subjects into state
-      if (this.state.subjects.length === 0 && summary) {
-        let subjects = []
-        for (let subject of summary.subjects) {
-          subjects.push({ label: 'sub-' + subject, value: subject })
-        }
-        subjects.reverse()
-        this.setState({ subjects })
+    // initialize subjects into state
+    if (this.state.subjects.length === 0 && this.props.dataset.summary) {
+      let subjects = []
+      for (let subject of this.props.dataset.summary.subjects) {
+        subjects.push({ label: 'sub-' + subject, value: subject })
       }
+      subjects.reverse()
+      this.setState({ subjects })
+    }
 
-      // pre-select snapshots
-      if (!this.state.selectedSnapshot) {
-        for (const snapshotIndex in snapshots) {
-          const snapshot = snapshots[snapshotIndex]
-          if (snapshot._id === dataset._id) {
-            if (snapshot.original) {
-              this._selectSnapshot({ target: { value: snapshot._id } })
-            } else if (snapshots.length > 1) {
-              this._selectSnapshot({
-                target: { value: snapshots[1]._id },
-              })
-            }
+    // pre-select snapshots
+    if (!this.state.selectedSnapshot) {
+      for (const snapshotIndex in this.props.snapshots) {
+        const snapshot = this.props.snapshots[snapshotIndex]
+        if (snapshot._id === this.props.dataset._id) {
+          if (snapshot.original) {
+            this._selectSnapshot({ target: { value: snapshot._id } })
+          } else if (this.props.snapshots.length > 1) {
+            this._selectSnapshot({
+              target: { value: this.props.snapshots[1]._id },
+            })
           }
         }
       }
@@ -81,48 +69,34 @@ class JobMenu extends Reflux.Component {
   }
 
   render() {
-    let datasets = this.state.datasets
-    let apps = datasets.apps
+    let apps = this.props.apps
     let selectedAppKey = this.state.selectedAppKey
     let selectedVersionID = this.state.selectedVersionID
-    let loading = datasets && datasets.loading
-    let loadingText = datasets.loadingApps
+    let loadingText = this.props.loadingApps
       ? 'Loading pipelines'
       : 'Starting ' + selectedVersionID
-
-    let returnLink = null
-    if (this.state.datasets.datasetUrl) {
-      returnLink = (
-        <Link to={this.state.datasets.datasetUrl}>
-          <button className="btn-reset">Back</button>
-        </Link>
-      )
-    }
-
     let form = (
-      <div className=" analysis-modal dataset-form-body col-xs-12">
-        <div className="dataset-form-content col-xs-12">
-          {this._snapshots()}
-          {this._apps()}
-          {selectedAppKey && selectedVersionID ? (
-            <div>
-              <Description
-                jobDefinition={apps[selectedAppKey][selectedVersionID]}
-              />
-              <Parameters
-                parameters={this.state.parameters}
-                parametersMetadata={this.state.parametersMetadata}
-                subjects={this.state.subjects}
-                onChange={this._updateParameter.bind(this)}
-                onRestoreDefaults={this._restoreDefaultParameters.bind(this)}
-              />
-              <span className="submit-warning">{this.state.submitWarning}</span>
-            </div>
-          ) : (
-            ''
-          )}
-          {this._submit()}
-        </div>
+      <div className="analysis-modal clearfix">
+        {this._snapshots()}
+        {this._apps()}
+        {selectedAppKey && selectedVersionID ? (
+          <div>
+            <Description
+              jobDefinition={apps[selectedAppKey][selectedVersionID]}
+            />
+            <Parameters
+              parameters={this.state.parameters}
+              parametersMetadata={this.state.parametersMetadata}
+              subjects={this.state.subjects}
+              onChange={this._updateParameter.bind(this)}
+              onRestoreDefaults={this._restoreDefaultParameters.bind(this)}
+            />
+            <span className="submit-warning">{this.state.submitWarning}</span>
+            {this._submit()}
+          </div>
+        ) : (
+          ''
+        )}
       </div>
     )
 
@@ -132,12 +106,14 @@ class JobMenu extends Reflux.Component {
           {this.state.error ? <h4 className="danger">Error</h4> : null}
           <h5>{this.state.message}</h5>
         </div>
-        {returnLink}
+        <button className="btn-admin-blue" onClick={this._hide.bind(this)}>
+          OK
+        </button>
       </div>
     )
 
     let body
-    if (this.state.loading || datasets.loadingApps) {
+    if (this.state.loading || this.props.loadingApps) {
       body = <Spinner active={true} text={loadingText} />
     } else if (this.state.message) {
       body = message
@@ -145,30 +121,16 @@ class JobMenu extends Reflux.Component {
       body = form
     }
 
-    let content = (
-      <div className="dataset-form">
-        <div className="col-xs-12 dataset-form-header">
-          <div className="form-group">
-            <label>Run Analysis</label>
-          </div>
-          <hr className="modal-inner" />
-          {body}
-        </div>
-      </div>
-    )
-
     return (
-      <ErrorBoundary
-        message="The dataset has failed to load in time. Please check your network connection."
-        className="col-xs-12 dataset-inner dataset-route dataset-wrap inner-route light text-center">
-        {loading ? (
-          <Timeout timeout={20000}>
-            <Spinner active={true} text={loadingText} />
-          </Timeout>
-        ) : (
-          content
-        )}
-      </ErrorBoundary>
+      <Modal show={this.props.show} onHide={this._hide.bind(this)}>
+        <Modal.Header closeButton>
+          <Modal.Title>Run Analysis</Modal.Title>
+        </Modal.Header>
+        <hr className="modal-inner" />
+        <Modal.Body>
+          <div className="dataset">{body}</div>
+        </Modal.Body>
+      </Modal>
     )
   }
 
@@ -181,7 +143,7 @@ class JobMenu extends Reflux.Component {
    * analysis application.
    */
   _apps() {
-    const apps = this.state.datasets ? this.state.datasets.apps : null
+    const apps = this.props.apps
     const selectedApp = this.state.selectedAppKey
 
     let validatedApps = batch.filterAppDefinitions(apps).reduce((acc, app) => {
@@ -258,13 +220,13 @@ class JobMenu extends Reflux.Component {
     )
 
     if (this.state.selectedSnapshot) {
-      const datasetName = this.state.datasets.dataset
-        ? this.state.datasets.dataset.name
-        : null
       return (
         <div>
           <hr />
-          <h5>Choose an analysis pipeline to run on dataset {datasetName}</h5>
+          <h5>
+            Choose an analysis pipeline to run on dataset{' '}
+            {this.props.dataset.name}
+          </h5>
           <div className="row">
             <div className="col-xs-12">
               <div className="col-xs-12 task-select">
@@ -317,12 +279,9 @@ class JobMenu extends Reflux.Component {
    * to run analysis on.
    */
   _snapshots() {
-    let datasets = this.state.datasets
-    let dataset = datasets ? datasets.dataset : null
-    let snapshots = datasets ? datasets.snapshots : null
     let options = []
-    if (snapshots) {
-      options = snapshots.map(snapshot => {
+    if (this.props.snapshots) {
+      options = this.props.snapshots.map(snapshot => {
         if (!snapshot.isOriginal && !snapshot.orphaned) {
           return (
             <option key={snapshot._id} value={snapshot._id}>
@@ -338,28 +297,16 @@ class JobMenu extends Reflux.Component {
     }
 
     let createSnapshot
-    if (dataset && dataset.access === 'admin') {
-      let dataset = datasetStore.data.dataset
-      let snapshots = datasetStore.data.snapshots
-      let modified = !(
-        snapshots.length > 1 &&
-        moment(dataset.modified).diff(moment(snapshots[1].modified)) <= 0
+    if (this.props.dataset.access === 'admin') {
+      createSnapshot = (
+        <div className="col-xs-6 default-reset">
+          <button
+            className="btn-reset"
+            onClick={this._createSnapshot.bind(this)}>
+            Create New Snapshot
+          </button>
+        </div>
       )
-      if (modified && datasets.datasetUrl) {
-        let to = {
-          pathname: this.state.datasets.datasetUrl + '/create-snapshot',
-          state: {
-            fromModal: 'jobs',
-          },
-        }
-        createSnapshot = (
-          <div className="col-xs-6 default-reset">
-            <Link to={to}>
-              <button className="btn-reset">Create New Snapshot</button>
-            </Link>
-          </div>
-        )
-      }
     }
 
     return (
@@ -388,33 +335,19 @@ class JobMenu extends Reflux.Component {
   }
 
   _submit() {
-    let selectedAppKey = this.state.selectedAppKey
-    let selectedVersionID = this.state.selectedVersionID
     if (this.state.disabledApps.hasOwnProperty(this.state.selectedVersion.id)) {
       return false
     }
-
-    let returnLink = this.state.datasets.datasetUrl ? (
-      <Link to={this.state.datasets.datasetUrl}>
-        <button className="btn-reset" onClick={this._hide.bind(this)}>
-          close
-        </button>
-      </Link>
-    ) : null
-
-    let startLink =
-      selectedAppKey && selectedVersionID ? (
+    return (
+      <div className="col-xs-12 modal-actions">
         <button
           className="btn-modal-submit"
           onClick={this._checkSubmitStatus.bind(this)}>
           Start
         </button>
-      ) : null
-
-    return (
-      <div className="col-xs-12 dataset-form-controls modal-actions">
-        {startLink}
-        {returnLink}
+        <button className="btn-reset" onClick={this._hide.bind(this)}>
+          close
+        </button>
       </div>
     )
   }
@@ -425,22 +358,22 @@ class JobMenu extends Reflux.Component {
    * Hide
    */
   _hide() {
-    // const success = this.state.message && !this.state.error
+    const success = this.state.message && !this.state.error
 
     // on modal close arguments
-    // const snapshotId = this.state.selectedSnapshot,
-    //   appLabel = this.state.selectedAppKey,
-    //   appVersion = this.state.selectedVersionID,
-    //   jobId = this.state.jobId
+    const snapshotId = this.state.selectedSnapshot,
+      appLabel = this.state.selectedAppKey,
+      appVersion = this.state.selectedVersionID,
+      jobId = this.state.jobId
 
-    // this.props.onHide(
-    //   success,
-    //   snapshotId,
-    //   appLabel,
-    //   appVersion,
-    //   jobId,
-    //   this.props.history,
-    // )
+    this.props.onHide(
+      success,
+      snapshotId,
+      appLabel,
+      appVersion,
+      jobId,
+      this.props.history,
+    )
     this.setState({
       loading: false,
       jobId: null,
@@ -500,9 +433,8 @@ class JobMenu extends Reflux.Component {
       }
       return !!requiredParameters[param]
     })
-    this.setState({ submitActive, submitWarning }, () => {
-      this.state.submitActive === true ? this._startJob() : null
-    })
+    this.setState({ submitActive, submitWarning })
+    this.state.submitActive === true ? this._startJob() : null
   }
 
   /**
@@ -548,10 +480,7 @@ class JobMenu extends Reflux.Component {
    */
   _selectApp(e) {
     const selectedAppKey = e.target.value
-    const selectedApp =
-      this.state.datasets && this.state.datasets.apps
-        ? this.state.datasets.apps[selectedAppKey]
-        : null
+    const selectedApp = this.props.apps[selectedAppKey]
     if (this.state.selectedAppKey !== e.target.value) {
       this.setState({
         parameters: [],
@@ -570,9 +499,9 @@ class JobMenu extends Reflux.Component {
    */
   _selectAppVersion(e) {
     const selectedVersionID = e.target.value
-    const selectedDefinition = this.state.datasets.apps[
-      this.state.selectedAppKey
-    ][selectedVersionID]
+    const selectedDefinition = this.props.apps[this.state.selectedAppKey][
+      selectedVersionID
+    ]
     const parametersMetadata = JSON.parse(
       JSON.stringify(selectedDefinition.parametersMetadata),
     )
@@ -634,7 +563,7 @@ class JobMenu extends Reflux.Component {
    */
   _startJob() {
     const selectedSnapshot = this.state.selectedSnapshot
-    const definitions = this.state.datasets.apps
+    const definitions = this.props.apps
     const key = this.state.selectedAppKey
     const revision = this.state.selectedVersionID
     const jobDefinition = definitions[key][revision]
@@ -654,16 +583,15 @@ class JobMenu extends Reflux.Component {
             let message, error
             if (err) {
               error = true
-              let response = err.response
-              if (err.status === 409) {
+              if (res.status === 409) {
                 message =
                   'This analysis has already been run on this dataset with the same parameters. You can view the results in the Analyses section of the dataset page.'
-              } else if (err.status === 503) {
+              } else if (res.status === 503) {
                 message =
                   'We are temporarily unable to process this analysis. Please try again later. If this issue persists, please contact the site administrator.'
-              } else if (err.status === 403 && response.body.error) {
+              } else if (res.status === 403 && res.body.error) {
                 // If non admins try to run more than 2 jobs at a time, want to display message letting them know they don't have access
-                message = response.body.error
+                message = res.body.error
               } else {
                 message =
                   'There was an issue submitting your analysis. Double check your inputs and try again. If the issue persists, please contact the site administrator.'
@@ -673,7 +601,7 @@ class JobMenu extends Reflux.Component {
                 'Your analysis has been submitted. You will receive a notification by email once the job is complete.'
             }
 
-            if (res && res.body && 'jobId' in res.body && res.body.jobId) {
+            if ('jobId' in res.body && res.body.jobId) {
               this.setState({
                 loading: false,
                 message: message,
@@ -698,10 +626,10 @@ JobMenu.propTypes = {
   apps: PropTypes.object,
   dataset: PropTypes.object,
   loadingApps: PropTypes.bool,
+  onHide: PropTypes.func.isRequired,
   show: PropTypes.bool,
   snapshots: PropTypes.array,
   history: PropTypes.object,
-  location: PropTypes.object,
 }
 
 JobMenu.defaultProps = {
