+++ app/src/scripts/dataset/tools/subscribe.jsx
@@ -3,29 +3,25 @@
 // dependencies -------------------------------------------------------
 
 import React from 'react'
-import Reflux from 'reflux'
 import PropTypes from 'prop-types'
 import userStore from '../../user/user.store'
 import userActions from '../../user/user.actions'
 import datasetActions from '../../dataset/dataset.actions'
-import datasetStore from '../dataset.store.js'
-import { Link, Redirect, withRouter } from 'react-router-dom'
-import { refluxConnect } from '../../utils/reflux'
+import { Modal } from '../../utils/modal.jsx'
 
-class Subscribe extends Reflux.Component {
+export default class Subscribe extends React.Component {
   // life cycle events --------------------------------------------------
 
   constructor(props) {
     super(props)
-    refluxConnect(this, datasetStore, 'datasets')
     this.state = {
+      subscribed: this.props.subscribed,
       hasToken: userStore.hasToken(),
       loading: false,
     }
-    this._createSubscription = datasetStore.createSubscription.bind(
-      this,
-      () => {},
-    )
+    this._createSubscription = this.props.createSubscription.bind(this, () => {
+      this.props.onHide()
+    })
   }
 
   componentWillReceiveProps() {
@@ -41,32 +37,25 @@ class Subscribe extends Reflux.Component {
   }
 
   render() {
-    if (
-      this.state.hasToken &&
-      this.state.datasets &&
-      this.state.datasets.dataset &&
-      this.state.datasets.dataset.subscribed &&
-      this.state.datasets.datasetUrl
-    ) {
-      return <Redirect to={this.state.datasets.datasetUrl} />
-    }
     return (
-      <div className="dataset-form">
-        <div className="col-xs-12 dataset-form-header">
-          <div className="form-group">
-            <label>Follow Dataset</label>
+      <Modal
+        show={this.props.show}
+        onHide={this.props.onHide}
+        className="download-modal">
+        <Modal.Header closeButton>
+          <Modal.Title>Follow Dataset</Modal.Title>
+        </Modal.Header>
+        <hr className="modal-inner" />
+        <Modal.Body>
+          <div className="dataset">
+            {this._message()}
+            <div>
+              <span className="caret-down" />
+              {this._controls()}
+            </div>
           </div>
-          <hr className="modal-inner" />
-        </div>
-        <div className="dataset-form-body col-xs-12">
-          <div className="dataset-form-content col-xs-12">
-            <div className="dataset share-modal">{this._message()}</div>
-          </div>
-          <div className="dataset-form-controls col-xs-12">
-            {this._controls()}
-          </div>
-        </div>
-      </div>
+        </Modal.Body>
+      </Modal>
     )
   }
 
@@ -93,25 +82,19 @@ class Subscribe extends Reflux.Component {
   }
 
   _followButton() {
-    if (
-      this.state.hasToken &&
-      !this.state.datasets.subscribed &&
-      this.state.datasets.datasetUrl
-    ) {
+    if (this.state.hasToken && !this.props.subscribed) {
       return (
-        <Link to={this.state.datasets.datasetUrl}>
-          <button
-            className="btn-modal-submit"
-            onClick={this._createSubscription.bind(this)}>
-            follow
-          </button>
-        </Link>
+        <button
+          className="btn-modal-submit"
+          onClick={this._createSubscription.bind(this)}>
+          follow
+        </button>
       )
     }
   }
 
   _followText() {
-    if (!this.state.datasets.subscribed) {
+    if (!this.props.subscribed) {
       return (
         <h5>
           If you would like to stay up to date about dataset updates and
@@ -122,19 +105,17 @@ class Subscribe extends Reflux.Component {
   }
 
   _continueButton() {
-    if (this.state.hasToken && this.state.datasets.subscribed) {
+    if (this.state.hasToken && this.props.subscribed) {
       return (
-        <Link to={this.props.location.pathname}>
-          <button className="btn-modal-submit" onClick={this.props.onHide}>
-            got it
-          </button>
-        </Link>
+        <button className="btn-modal-submit" onClick={this.props.onHide}>
+          got it
+        </button>
       )
     }
   }
 
   _continueText() {
-    if (this.state.hasToken && this.state.datasets.subscribed) {
+    if (this.state.hasToken && this.props.subscribed) {
       return (
         <h5>
           You are already following this dataset. You will continue to receive
@@ -157,13 +138,11 @@ class Subscribe extends Reflux.Component {
   }
 
   _returnButton() {
-    if (!this.state.datasets.subscribed && this.state.datasets.datasetUrl) {
+    if (!this.props.subscribed) {
       return (
-        <Link to={this.state.datasets.datasetUrl}>
-          <button className="btn-reset" onClick={this.props.onHide}>
-            no, thanks
-          </button>
-        </Link>
+        <button className="btn-reset" onClick={this.props.onHide}>
+          no, thanks
+        </button>
       )
     }
   }
@@ -174,8 +153,4 @@ Subscribe.propTypes = {
   show: PropTypes.bool,
   onHide: PropTypes.func,
   createSubscription: PropTypes.func,
-  location: PropTypes.object,
-  history: PropTypes.object,
 }
-
-export default withRouter(Subscribe)
