+++ server/libs/email/templates/snapshot-created.html
@@ -45,9 +45,6 @@
 		border: 1px solid #ccc;
 		padding: 15px;
 	}
-	.changelog {
-		white-space: pre-wrap;
-	}
 	.FAILED {color: #d9534f;}
 	.FINISHED {color: #5cb85c;}
 </style>
@@ -65,11 +62,6 @@
 
 		<div>
 			<a class='dataset-link' href="{{siteUrl}}/datasets/{{datasetId}}/versions/{{versionNumber}}">Click here to view this snapshot on OpenNeuro &raquo;</a>
-		</div>
-
-		<div>
-			<h4>Changelog</h4>
-			<p class="changelog">{{changelog}}</p>
 		</div>
 
 		<p>
